moved events to own file again and README
