+++ README.md
@@ -2,10 +2,13 @@
 
 This is the repository for the code which accompanies my post on Event Sourcing in Go. The original post is [here](http://jen20.com/2015/02/08/event-sourcing-in-go.html).
 
-Remade by @hallgren
+Remade by [@hallgren](https://github.com/hallgren)
 
 To build and run:
 
-- `go build`
-- `./go-event-sourcing-sample`
+- `make`
+- `./goes`
 
+To clean generated code and the binary:
+
+- `make clean`
