+++ frequent_flier_account.go
@@ -14,35 +14,6 @@ type FrequentFlierAccount struct {
 	status        Status
 }
 
-// Status represents the Red, Silver or Gold tier level of a FrequentFlierAccount
-type Status int
-
-const (
-	StatusRed    Status = iota
-	StatusSilver Status = iota
-	StatusGold   Status = iota
-)
-
-// go:generate stringer -type=Status
-
-type FrequentFlierAccountCreated struct {
-	AccountId         string
-	OpeningMiles      int
-	OpeningTierPoints int
-}
-
-type StatusMatched struct {
-	NewStatus Status
-}
-
-type FlightTaken struct {
-	MilesAdded      int
-	TierPointsAdded int
-}
-
-type PromotedToGoldStatus struct {
-}
-
 // CreateFrequentFlierAccount constructor
 func CreateFrequentFlierAccount(id string) *FrequentFlierAccount {
 	self := &FrequentFlierAccount{}
