+++ frequent_flier_events.go
@@ -1,0 +1,29 @@
+package main
+
+// Status represents the Red, Silver or Gold tier level of a FrequentFlierAccount
+type Status int
+
+const (
+	StatusRed    Status = iota
+	StatusSilver Status = iota
+	StatusGold   Status = iota
+)
+
+// go:generate stringer -type=Status
+
+type FrequentFlierAccountCreated struct {
+	AccountId         string
+	OpeningMiles      int
+	OpeningTierPoints int
+}
+
+type StatusMatched struct {
+	NewStatus Status
+}
+
+type FlightTaken struct {
+	MilesAdded      int
+	TierPointsAdded int
+}
+
+type PromotedToGoldStatus struct{}
